refactor: decompoe PrescricaoMedica em View e subcomponentes
- Substitui o componente monolítico PrescricaoMedica.vue por PrescricaoView.vue
- Extrai seções de UI para componentes dedicados: Identificacao, Protocolo, PreQt, Qt, PosQt e Footer
- Atualiza o router para apontar para a nova View
- Centraliza a lógica de estado e API na View, delegando a renderização para os filhos

Fixes #9
